switch recommendation to hex instead of base64 for passwords.
Base64 can sometimes have non-URL safe characters in it, which is no good for
an application that logs in to its database using basic auth in the URL. Using
hex instead will only use numbers and letters.

I'm unsure if this is the right recommendation, since it significantly reduces
the total possible password combinations. I think it's the most user-friendly,
though.